Updated README.md
Testing indicates the webcrawler script has a small bug regarding some records having mismatched listing urls and listing ids. The README now has a section to take note of this or any other existing bugs in this webcrawler program. While this bug is worthy of note, it will not affect the validity or accuracy of 44 of the 46 total columns, including each of the columns that one would want to use in a regression or data analysis of rental data.