Update README.md
Removed "Current Shortcomings and Bugs" section since the webcrawler misaligned scraped data bug has been fixed. 

To achieve this, I needed to revise the try...finally clause--ie, nested within the for loop that iterates over the rental listing URL pages--to be a try except else finally clause. The try except statements check for whether a rental listing's URL has been deleted or not. If a listing has been deleted, then the except clause triggers, and we append 'nan'  values to each of the scraped data lists--ie, ids, prices, cities, etc. Why? Because we need to ensure the scraped data are not misaligned by adding in some value to line up with the listing_urls list.

For any rental listing that still exists--ie, has not been deleted--then the else statement is triggered. I needed to revise the parse_html_via_xpath() functions that parse each HTML object. Namely: I needed to add nested try except clauses to check whether--for a given rental listing--data actually exists for a given HTML element that we are trying to scrape. The nested except clause will trigger whenever a given HTML element does not exist for a given rental listing. Namely, the script will append an 'nan' value to account for missing data for the given list with respect to that given HTML element.

Then, the finally clause will print out some of the scraped data as a sanity check. I also included an if else statement in the finally clause to ensure the webcrawler will have an exit condition since the for loop and try except else finally statements are embedded within a while loop. Basically, the finally clause will check for the presence of any duplicate listing ids, and immediately terminate the for loop once any duplicate ids are detected.