Specify xpath arguments as lists and join pipe operator so we can avoid hardcoding these arguments
Some of the methods from the webcrawler's Craigslist_Rentals object rely on parsing multiple xpaths (ie, one of multiple xpaths are possible as the server has been updated and changed) as one (or more) of the methods' arguments. For instance, xpaths_listing_urls--which refers to each of the possible xpaths pertaining to the href URLs of the rental listing pages--is an argument of the obtain_listing_urls() method, which uses selenium's find_element('xpath') method to locate HTML elements based on a given xpath.  

In order to have selenium's find_element() method look up more than one potential xpaths via a Boolean "or" condition, we can supply the find_element() method with *multiple* possible xpaths that are each separated by the pipe (ie, "|") operator.

For any such arguments--ie, with multiple possible xpaths--these have been transformed to be lists instead of being written as individual strings. However, selenium's find_element() method needs a single *string* as its argument (ie, we cannot use a list directly). Thus, we transform *each element* from this list *into* a single string by using f strings within a list comprehension, and apply the .join() method to each element of the list so that each element is separated by pipe operators. 

This way, we can avoid hardcoding these xpath arguments. As a result, we can thereby make the code easier to read and any future changes to the xpaths (based on changes to craigslist's server) easier to implement.