Reference current path instead of manually typing
In main.py executable, use getcwd() method from os library to reference the current working directory (path) instead of manually typing the full path.

Why?: To avoid other users--whose drive or parent path names (ie, the parent path relative to this project's scripts saved onto the local machine) might differ--from having to manually changing the path to their local machine's path. 

By referencing the current working directory, we ensure the path never needs to be changed when using the scripts on a different machine or used by a different user. The reason for this is the main.py can *only* be executed *after* a user has changed the working directory to the parent folder of this webcrawler project scripts saved on the given local machine.