Update README
Corrected SQL query example under "Possible Use Cases" section. Namely: the 10 cheapest such apartments from SF for January 2023 query example. Added PARTITION BY clause to DENSE_RANK() in common table expression since we needed to refine the window that the DENSE_RANK() is applying to in order to consider each city as a window, not the entire SF bay data as a whole.