Create SQL_create_database_clist.py
Create new data pipeline file that creates a new SQL Server database if it does not already exist. This way, the user does not have to manually create a database using the SQL Server SSMS GUI or a SQL Server command-line.

This database will be used to store the scraped rental listing data.