Add new Python script
Upload new Python script to root directory. This script will print the latest modified files for each subdirectory within the scraped_data folder. This script's purpose is to show the user when the webcrawler was last run for a given craigslist region and subregion. Therefore, the user can more efficiently decide which region or subregion to run the webcrawler on next, given for example that certain regions or subregions would have less recent data.