Update selenium_webcrawler.py
Improve next page navigation by clicking on next page until the next page no longer exists. This is verified by including a new if statement--defined as a new next_page_element object--that will check if the element does not equal 0, and if this is true then keep clicking that next page button/element.