Update Pandas_and_SQL_ETL_and_data_cleaning.py
Correct code logic and control flow to enable regions--that have never been stored in the SQL table before--to be able to bypass the date-time filters that would otherwise apply (i.e., be applied to regions that already have data stored in the SQL table). 

The reason  behind this is simple: there would need to be at least some data stored in the SQL table for the given region for there to be any value for the date_posted data pertaining to the given region!