Update main.py
Update main to include a working xpath reference to the subregion codes that we need for parsing subregions for caraigslist regions (such as Chicago, IL; San Diego, CA, etc.) that have subregions. More specifically, this is in regards to the ul_subregions_xpath argument for the parse_subregions_via_xpath() method call. 

To be clear, the original argument referenced no longer extant HTML of craigslist's server, and I am unsure of when the change in the HTML source code occurred.