Update determine_subregions_for_given_clist_region.py
Update script to a) allow for Chrome webdriver to install correctly by using updated selenium library syntax. Also: b) update reference to subregion codes to match that of the argument given in main.py, so that we can correctly parse subregion codes for regions (other than those we already have for the SF Bay Area) that have sub-regions.