Update README.md
Add new item in To-Dos heading list. Namely: mention that I might want to make a more sophisticated drop-down list that will prompt the user to first select the country, then state or province, and finally region they want the webcrawler to run for. 
a hierarchical list for the region selection drop-downs that are